Implementation of ble connection for iOS (#179)

+++ ios/Classes/DoozMeshNetwork/Channels/Arguments/GetGroupElementIdsArguments.swift
@@ -6,5 +6,5 @@
 //
 
 struct GetGroupElementIdsArguments: BaseFlutterArguments {
-    let groupAddress: Int16
+    let groupAddress: Int
 }
